requirements.txt
absl-py==2.1.0
accelerate==0.33.0
aiohappyeyeballs==2.3.4
aiohttp==3.10.0
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.7.0
asgiref==3.8.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
audioread==3.0.1
autobahn==24.4.2
Automat==22.10.0
autopep8==2.3.1
beautifulsoup4==4.12.3
bertopic==0.16.3
billiard==4.2.0
bioc==2.1
bleach==6.1.0
blis==0.7.11
boto3==1.35.0
botocore==1.35.0
catalogue==2.0.10
celery==5.4.0
certifi==2024.7.4
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudpathlib==0.18.1
cloudpickle==2.2.1
confection==0.1.5
conllu==4.5.3
constantly==23.10.4
contourpy==1.2.1
crispy-bootstrap5==2024.2
cron-descriptor==1.4.3
cryptography==43.0.0
cycler==0.12.1
cymem==2.0.8
daphne==4.1.2
datasets==2.21.0
decorator==5.1.1
Deprecated==1.2.14
dill==0.3.8
Django==5.0.7
django-allauth==64.0.0
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-cleanup==8.1.0
django-cors-headers==4.4.0
django-crispy-forms==2.3
django-debug-toolbar==4.4.6
django-filter==24.3
django-jazzmin==3.0.0
django-timezone-field==7.0
djangorestframework==3.15.2
docker==7.1.0
docopt==0.6.2
drf-spectacular==0.27.2
filelock==3.15.4
flair==0.14.0
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
ftfy==6.2.3
gast==0.6.0
gdown==5.2.0
gensim==4.3.3
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.65.5
gTTS==2.5.3
gunicorn==22.0.0
h5py==3.11.0
hdbscan==0.8.38.post1
huggingface-hub==0.24.5
hyperlink==21.0.0
idna==3.7
importlib-metadata==6.11.0
incremental==24.7.2
inflection==0.5.1
intervaltree==3.1.0
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonlines==4.0.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
keras==3.5.0
kiwisolver==1.4.5
kombu==5.3.7
langchain==0.2.14
langchain-core==0.2.33
langchain-huggingface==0.0.3
langchain-text-splitters==0.2.2
langcodes==3.4.0
langdetect==1.0.9
langsmith==0.1.96
language_data==1.2.0
latex2mathml==3.77.0
lazy_loader==0.4
libclang==18.1.1
librosa==0.10.2.post1
llvmlite==0.43.0
lxml==5.3.0
marisa-trie==1.2.0
Markdown==3.6
markdown-it-py==3.0.0
markdown2==2.5.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
ml-dtypes==0.4.0
more-itertools==10.4.0
mpld3==0.5.10
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
multiprocess==0.70.16
murmurhash==1.0.10
namex==0.0.8
networkx==3.3
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
opencv-python==4.10.0.84
opt-einsum==3.3.0
optree==0.12.1
orjson==3.10.6
packaging==24.1
pandas==2.2.2
pathos==0.3.2
pgvector==0.3.2
pillow==10.4.0
platformdirs==4.2.2
plotly==5.23.0
pooch==1.8.2
pox==0.3.4
ppft==1.7.6.8
pptree==3.1
preshed==3.0.9
prompt_toolkit==3.0.47
protobuf==4.25.4
psutil==6.0.0
psycopg==3.2.1
psycopg-binary==3.2.1
psycopg-pool==3.2.2
py-cpuinfo==9.0.0
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
PyJWT==2.9.0
pynndescent==0.5.13
pyOpenSSL==24.2.1
pyparsing==3.1.2
PySocks==1.7.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytorch_revgrad==0.2.0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.8
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
rpds-py==0.20.0
s3transfer==0.10.2
safetensors==0.4.4
sagemaker==2.229.0
schema==0.7.7
scikit-learn==1.5.1
scipy==1.13.1
seaborn==0.13.2
segtok==1.5.11
semver==3.0.2
sentence-transformers==3.0.1
sentencepiece==0.2.0
service-identity==24.1.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
smdebug-rulesconfig==1.0.1
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.6
soxr==0.4.0
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.10.4
SQLAlchemy==2.0.31
sqlitedict==2.1.0
sqlparse==0.5.1
srsly==2.4.8
svgwrite==1.4.3
sympy==1.13.2
tabulate==0.9.0
tblib==3.0.0
tenacity==8.5.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow-hub==0.16.1
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-text==2.17.0
termcolor==2.4.0
tf_keras==2.17.0
thinc==8.2.5
threadpoolctl==3.5.0
tokenizers==0.19.1
tomli==2.0.1
torch==2.4.0
torchvision==0.19.0
tqdm==4.66.5
transformer-smaller-training-vocab==0.4.0
transformers==4.44.0
triton==3.0.0
Twisted==24.3.0
txaio==23.1.1
typer==0.12.4
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.2.79
ultralytics-thop==2.0.5
umap-learn==0.5.6
uritemplate==4.1.1
urllib3==2.2.2
vine==5.1.0
wasabi==1.1.3
wavedrom==2.0.3.post3
wcwidth==0.2.13
weasel==0.4.1
webencodings==0.5.1
Werkzeug==3.0.3
Wikipedia-API==0.6.0
wrapt==1.16.0
xxhash==3.5.0
yarl==1.9.4
zipp==3.20.0
zope.interface==6.4.post2
